Move search icon from FAB to Appbar
