+++ app/src/main/java/com/devkproject/movieinfo2/ui/component/appbar/HomeAppBar.kt
@@ -3,7 +3,9 @@ package com.devkproject.movieinfo2.ui.component.appbar
 import androidx.compose.material.*
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Menu
+import androidx.compose.material.icons.filled.Search
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.tooling.preview.Preview
 import com.devkproject.movieinfo2.ui.theme.Purple500
 
@@ -23,6 +25,15 @@ fun HomeAppBar(title: String, openDrawer: () -> Unit, openFilters: () -> Unit) {
                 Icon(Icons.Default.Menu, contentDescription = "Menu")
             }
         },
+        actions = {
+            IconButton(onClick = openFilters) {
+                Icon(
+                    imageVector = Icons.Filled.Search,
+                    contentDescription = "Search",
+                    tint = Color.White,
+                )
+            }
+        }
     )
 }
 
@@ -31,3 +42,17 @@ fun HomeAppBar(title: String, openDrawer: () -> Unit, openFilters: () -> Unit) {
 fun HomeAppBarTest() {
     HomeAppBar(title = "MovieInfo", openDrawer = { }, openFilters = { })
 }
+
+//Spacer(modifier = Modifier.weight(4f))
+//
+//Image(
+//imageVector = Icons.Filled.Search,
+//colorFilter = ColorFilter.tint(Color.White),
+//contentDescription = null,
+//modifier = Modifier
+//.align(Alignment.CenterVertically)
+//.padding(end = 10.dp)
+//.clickable {
+//
+//}
+//)
